moved the parallel contingency analsysis in its own file. Also I've added a test for it.
